refactor(ui): TemizleModali'yi ayri dosyaya cikar + sabitler modulu (Faz E)
- screens/TakvimAyarlari/sabitler.ts: paylasilan VAKIT_* sabitleri + saatFormatla
- screens/TakvimAyarlari/TemizleModali.tsx: ~550 satirlik 4-asamali etkinlik
  temizleme modali (kriter/taraniyor/onay/tamamlandi)
- TakvimAyarlari 1449 -> 876 satir (bu dalda toplam 1786 -> 876, ~%51 azalma)

391 test geciyor, 0 tip hatasi.
